refactor(backend-test): refactor test
- to ensure each module performs a specific task

[Finishes #153083406]
